+++ src/Features/Core/Portable/FindSymbols/SymbolTree/MetadataInfo.cs
